+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 `unreleased`_
 -------------
 * Add `offline` option to `make_dropbox_blueprint`
+* Added Atlassian pre-set configuration
 
 `3.2.0`_ (2020-11-24)
 ---------------------